plugins/telegram/mcp-server/pyproject.toml
[project]
name = "telegram-mcp-server"
version = "0.1.0"
description = "MCP server for accessing Telegram conversations via natural language"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.5.0",
    "telethon>=1.36.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]
